Add a benchmark
This commit adds a benchmark using criterion framework.
To run the benchmark:

    cargo bench

The results in html form are in `target/criterion/report` directory.

This benchmarks this library against crates order-stat, kth and
pdqselect.

Currently, it benchmarks finding median of a randomly shuffled `Vec<u32>`
for sizes in the range 1k-1M.